requirements.txt
# 基因文献智能检索助手 - 依赖包

# 生物信息学
biopython>=1.81

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP 请求
requests>=2.31.0

# 配置文件
pyyaml>=6.0

# LLM API
openai>=1.0.0

# 语义向量化
sentence-transformers>=2.2.0

# 报告生成
markdown>=3.5.0
openpyxl>=3.1.0

# 可视化
matplotlib>=3.7.0
networkx>=3.0
plotly>=5.15.0

# 开发工具
pytest>=7.4.0